+++ relay-base-schema/src/data_category.rs
@@ -92,6 +92,15 @@ pub enum DataCategory {
     Uptime = 21,
     /// Counts the number of individual attachments, as opposed to the number of bytes in an attachment.
     AttachmentItem = 22,
+    /// LogItem
+    ///
+    /// This is the category for logs for which we store the count log events for users for measuring
+    /// missing breadcrumbs, and count of logs for rate limiting purposes.
+    LogItem = 23,
+    /// LogByte
+    ///
+    /// This is the category for logs for which we store log event total bytes for users.
+    LogByte = 24,
     //
     // IMPORTANT: After adding a new entry to DataCategory, go to the `relay-cabi` subfolder and run
     // `make header` to regenerate the C-binding. This allows using the data category from Python.
@@ -120,6 +129,8 @@ impl DataCategory {
             "transaction_indexed" => Self::TransactionIndexed,
             "monitor" => Self::Monitor,
             "span" => Self::Span,
+            "log_item" => Self::LogItem,
+            "log_byte" => Self::LogByte,
             "monitor_seat" => Self::MonitorSeat,
             "feedback" => Self::UserReportV2,
             "user_report_v2" => Self::UserReportV2,
@@ -152,6 +163,8 @@ impl DataCategory {
             Self::TransactionIndexed => "transaction_indexed",
             Self::Monitor => "monitor",
             Self::Span => "span",
+            Self::LogItem => "log_item",
+            Self::LogByte => "log_byte",
             Self::MonitorSeat => "monitor_seat",
             Self::UserReportV2 => "feedback",
             Self::MetricBucket => "metric_bucket",